bugfix:홈페이지 경로 수정
