+++ src/app/loading.tsx
@@ -1,1 +1,11 @@
 'use client';
+
+const LoadingPage = () => {
+  return (
+    <>
+
+    </>
+  );
+};
+
+export default LoadingPage;
