+++ src/app/page.tsx
@@ -1,31 +1,10 @@
-'use client';
-
-import { createClient } from '@supabase/supabase-js';
-import { useEffect, useState } from 'react';
-import UserHomePage from './my-page/components/home/UserHomePage';
-import GuestHomePage from './my-page/components/home/GuestHomepage';
-
-const supabase = createClient(process.env.NEXT_PUBLIC_SUPABASE_URL!, process.env.NEXT_PUBLIC_SUPABASE_ANON_KEY!);
+import ClientHomePage from "@/components/home/ClientHomePage";
 
 const HomePage = () => {
-  const [isLogin, setIsLogin] = useState(false);
-
-  useEffect(() => {
-    const ckeckLoginStatus = async () => {
-      const { data } = await supabase.auth.getSession();
-      setIsLogin(!!data.session);
-    };
-
-    ckeckLoginStatus();
-
-    const { data: listener } = supabase.auth.onAuthStateChange((_envent, session) => {
-      setIsLogin(!!session);
-    });
-
-    return () => {
-      listener.subscription.unsubscribe();
-    };
-  }, []);
-
-  return isLogin ? <UserHomePage /> : <GuestHomePage />;
+  return (
+    <div>
+      <ClientHomePage/>
+    </div>
+  );
 };
+export default HomePage;
